requirements.txt
django
MySQL-python
pip
ez_setup
ConfigParser
django-phonenumber-field
django-phonenumber-field[phonenumbers]
django-phonenumber-field[phonenumberslite]